refactor: clean up imports/file names, improve code quality (#29)
