+++ src/components/cards/Port.astro
@@ -1,6 +1,6 @@
 ---
 import type { ColorName } from "@catppuccin/palette";
-import type { Category } from "../../lib/getPorts";
+import type { Category } from "../../lib/ports";
 
 import Icon from "../icons/Ports.astro";
 import PillList from "../lists/Pills.astro";
