+++ src/components/icons/Ports.astro
@@ -1,6 +1,6 @@
 ---
 import type { ColorName } from "@catppuccin/palette";
-import { getIcon } from "../../lib/getIcon";
+import { getIcon } from "../../lib/icon";
 
 interface Props {
   port: string;
