+++ src/layouts/imports/Footer.astro
@@ -2,7 +2,7 @@
 import Logo from "../../components/icons/Logo.astro";
 import Link from "../../components/Link.astro";
 
-import { footer as navLinkList } from "../../contents/Navigations";
+import { footer as navLinkList } from "../../contents/navigation";
 ---
 
 <footer>
