+++ src/layouts/imports/Header.astro
@@ -1,7 +1,7 @@
 ---
 import Logo from "../../components/icons/Logo.astro";
 
-import { header as navLinkList } from "../../contents/Navigations";
+import { header as navLinkList } from "../../contents/navigation";
 ---
 
 <header>
