+++ src/pages/palette.astro
@@ -6,7 +6,7 @@ import PageIntro from "../components/PageIntro.astro";
 import PhosphorIcon from "../components/icons/Phosphor.astro";
 import FlavorName from "../components/FlavorName.astro";
 
-import { CatppuccinLinks as links } from "../contents/CatppuccinLinks";
+import { links } from "../contents/links";
 
 import { flavorEntries } from "@catppuccin/palette";
 ---
