+++ src/pages/ports.astro
@@ -1,5 +1,5 @@
 ---
-import { ports, categories, type Category } from "../lib/getPorts";
+import { ports, categories, type Category } from "../lib/ports";
 
 import Layout from "../layouts/Layout.astro";
